+++ .vscode/tasks.json
@@ -4,7 +4,7 @@
   "version": "0.1.0",
   "command": "npm",
   "isShellCommand": true,
-  "showOutput": "always",
+  "showOutput": "never",
   "suppressTaskName": true,
   "tasks": [
     {
