+++ src/app.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 import * as bodyParser from 'body-parser';
 import * as cors from 'cors';
 import { env } from './env.config';
-import { loginRouter } from './api'
+import { loginRouter } from './routes';
 
 const app = express();
 app.use(bodyParser.urlencoded({ extended: true }));
